+++ Nero/Param.hs
@@ -13,17 +13,17 @@ module Nero.Param
   -- * MultiMap
   , MultiMap
   , fromList
-  , encodeMultiMap
+  , null
   ) where
 
-import Data.ByteString.Lazy (ByteString)
-import qualified Data.ByteString.Lazy.Char8 as B8
+import Prelude hiding (null)
 import Data.Text.Lazy (Text, intercalate)
 import Data.Map (Map)
 import qualified Data.Map as Map
 import Data.Text.Lazy.Lens (utf8)
 
 import Nero.Prelude
+import Nero.Binary
 
 -- * HTTP Parameters
 
@@ -36,10 +36,7 @@ class Param a where
 -- | A 'Map' with multiple values. Also known as a @MultiDict@ in most web
 --   frameworks.
 newtype MultiMap = MultiMap { unMultiMap :: Map Text [Text] }
-                   deriving (Eq)
-
-instance Show MultiMap where
-    show = B8.unpack . encodeMultiMap
+                   deriving (Eq, Show)
 
 -- | The default monoid implementation is left biased, this implementation
 --   /mappends/ the values.
@@ -63,17 +60,20 @@ instance At MultiMap where
 instance Param MultiMap where
     param k = ix k . traverse
 
-fromList :: [(Text, [Text])] -> MultiMap
-fromList = MultiMap . Map.fromListWith (++)
-
 -- | Encode a 'MultiMap' with the typical query string format. This is
 --   useful to render 'MultiMap's when testing. The web server adapter for
 --   @Nero@ should do this for you in the real application.
-encodeMultiMap :: MultiMap -> ByteString
-encodeMultiMap =
-    review utf8
-  . intercalate "&"
-  -- Map.foldMapWithKey not supported in `containers-0.5.0.0` coming with
-  -- GHC==7.6.3
-  . fold . Map.mapWithKey (map . mappend . flip mappend "=")
-  . unMultiMap
+
+instance Renderable MultiMap where
+    render = review utf8
+           . intercalate "&"
+           -- Map.foldMapWithKey not supported in `containers-0.5.0.0` coming with
+           -- GHC==7.6.3
+           . fold . Map.mapWithKey (map . mappend . flip mappend "=")
+           . unMultiMap
+
+fromList :: [(Text, [Text])] -> MultiMap
+fromList = MultiMap . Map.fromListWith (++)
+
+null :: MultiMap -> Bool
+null = Map.null . unMultiMap